Reorganize project tree
