+++ app/src/main/java/com/example/padamlight/enums/MarkerType.java
@@ -1,0 +1,9 @@
+package com.example.padamlight.enums;
+
+/**
+ * Market enum
+ * Define if marker is pickup type or dropoff type
+ */
+public enum MarkerType {
+    PICKUP, DROPOFF;
+}
