+++ app/src/main/java/com/example/padamlight/ui/base/BaseFragment.java
@@ -1,0 +1,50 @@
+package com.example.padamlight.ui.base;
+
+
+import android.os.Bundle;
+
+
+import android.support.annotation.LayoutRes;
+import android.support.annotation.NonNull;
+import android.support.annotation.Nullable;
+import android.support.v4.app.Fragment;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+public abstract class BaseFragment extends Fragment {
+
+    @LayoutRes
+    protected int layoutRes;
+
+
+    public BaseFragment(@LayoutRes int layoutRes){
+        this.layoutRes = layoutRes;
+    }
+
+    @Override
+    public void onCreate(@Nullable Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+
+    }
+
+    @Nullable
+    @Override
+    public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
+        super.onCreateView(inflater, container, savedInstanceState);
+        return inflater.inflate(layoutRes, container, false);
+
+    }
+
+    @Override
+    public void onViewCreated(@NonNull View view, @Nullable Bundle savedInstanceState) {
+        super.onViewCreated(view, savedInstanceState);
+        initUi();
+    }
+
+
+    public abstract void initUi();
+
+
+
+}
