+++ app/src/main/java/com/example/padamlight/ui/map/interfaces/MapActionsDelegate.java
@@ -1,0 +1,16 @@
+package com.example.padamlight.ui.map.interfaces;
+
+import com.example.padamlight.enums.MarkerType;
+import com.google.android.gms.maps.model.LatLng;
+
+/**
+ * Map interface
+ * Implement this in your page where there is a map to use map methods
+ */
+public interface MapActionsDelegate {
+    void updateMap(LatLng... latLngs);
+
+    void updateMarker(MarkerType type, String markerName, LatLng markerLatLng);
+
+    void clearMap();
+}
